+++ summawise/main.py
@@ -8,9 +8,12 @@ CONTEXT_SETTINGS = {"max_content_width": sys.maxsize}
 
 @click.group(context_settings = CONTEXT_SETTINGS)
 @click.version_option(str(VERSION), "-v", "--version")
-def cli():
+@click.option("--debug", is_flag = True, hidden = True, help = "Enable debug mode")
+@click.pass_context
+def cli(ctx: click.Context, debug: bool):
     """A tool to use AI to analyze and interact with vectorized data from custom sources of information."""
-    pass
+    ctx.ensure_object(dict)
+    ctx.obj["DEBUG"] = debug
 
 def register_commands():
     cli.add_command(scan)